Fix clippy warnings on x86_64 target
- Fixed simple lint warnings (single_match, borrowed_box, len_zero, while_immutable_condition, redundant_locals, explicit_counter_loop, Option.and_then, needlessly_taken_reference, manual_div_ceil, trim_split_whitespace, get_first, sort_by_key, unnecessary_cast, last() on DoubleEndedIterator, manual_map, needless_borrow, From preference, misnamed_getters, reference immediately dereferenced, legacy numeric constants, redundant_closure, enumerate with discarded index, collapsible_match, manual_is_multiple_of, needless_range_loop)
- Added #[allow(clippy::too_many_arguments)] to functions with many parameters
- Added #[allow(clippy::type_complexity)] to functions with complex document types
- Added #[allow(clippy::only_used_in_recursion)] to recursive function parameters
- Reduced clippy errors from 68 to 0 on x86_64 target

+++ crates/core/src/document/mupdf/page.rs
@@ -46,8 +46,8 @@ impl Page {
     pub fn dims(&self) -> (f32, f32) {
         let bounds = self.bound();
         (
-            (bounds.x1 - bounds.x0) as f32,
-            (bounds.y1 - bounds.y0) as f32,
+            bounds.x1 - bounds.x0,
+            bounds.y1 - bounds.y0,
         )
     }
