+++ crates/core/src/view/calculator/display.rs
@@ -180,7 +180,7 @@ impl Calculator {
 
         let kb_rect = rect![
             rect.min.x,
-            rect.max.y - (small_height + 3 * big_height) as i32 + big_thickness,
+            rect.max.y - (small_height + 3 * big_height) + big_thickness,
             rect.max.x,
             rect.max.y - small_height - small_thickness
         ];
